+++ contracts/rust/helloworld/src/contract.rs
@@ -22,7 +22,7 @@ impl HelloWorldFunc {
         HelloWorldFunc { sc: ScContractFunc::new(ctx, HSC_NAME) }
     }
 
-    pub fn delay(&mut self, seconds: i64) -> &mut HelloWorldFunc {
+    pub fn delay(&mut self, seconds: i32) -> &mut HelloWorldFunc {
         self.sc.delay(seconds);
         self
     }
